refactors example options and remove routing_first
